04-UT | Create Unit test about spliting rows from std::string to std::vector<std::string>
