--- lib/lib.h
@@ -1,7 +1,0 @@
-#pragma once
-
-#define REGEX_CONST "^((25[0-5]|2[0-4][\\d]|1[\\d][\\d]|[\\d]?[\\d])[\\.,\\s]){3}(25[0-5]|2[0-4][\\d]|1[\\d][\\d]|[\\d]?[\\d])$"
-
-int returnVersion();
-
-int isRegexpValueCorrect();
